refactor(common): move ErrorResponseBuilder and ErrorResponseDto to common layer
Shared error response utilities were scoped to catalog-policy but are generic
enough for any module. Moving to common/ prevents cross-module imports and
follows the established pattern (db-error.utils, pagination.dto). Old locations
re-export to preserve backward compatibility.
