+++ apps/api/src/common/dtos/index.ts
@@ -5,3 +5,4 @@ export * from './image.dto';
 export * from './external-ratings.dto';
 export * from './ratingo-stats.dto';
 export * from './pagination.dto';
+export * from './error-response.dto';
