+++ apps/api/src/common/utils/error-response.builder.spec.ts
