+++ apps/api/src/modules/catalog-policy/presentation/utils/error-response.builder.ts
@@ -1,56 +1,4 @@
 /**
- * Error Response Builder
- *
- * Utility for building standardized error responses.
+ * Re-export from common layer.
  */
-
-import { type ErrorResponseDto } from '../dto/error-response.dto';
-
-interface ErrorWithCode {
-  code: string;
-  message: string;
-}
-
-/**
- * Builds standardized error response objects.
- */
-export const ErrorResponseBuilder = {
-  /**
-   * Build error response from an error with code.
-   */
-  build(
-    error: ErrorWithCode,
-    statusCode: number,
-    details?: Record<string, unknown>,
-  ): ErrorResponseDto {
-    return {
-      success: false,
-      error: {
-        code: error.code,
-        message: error.message,
-        statusCode,
-        ...(details && { details }),
-      },
-    };
-  },
-
-  /**
-   * Build error response with custom code (for errors without code property).
-   */
-  buildWithCode(
-    code: string,
-    message: string,
-    statusCode: number,
-    details?: Record<string, unknown>,
-  ): ErrorResponseDto {
-    return {
-      success: false,
-      error: {
-        code,
-        message,
-        statusCode,
-        ...(details && { details }),
-      },
-    };
-  },
-};
+export { ErrorResponseBuilder } from '../../../../common/utils/error-response.builder';
